support linking to multiple vers of a dylib
Change relink() to check that the library being linked, and the link
path in the target share the same dylib basename prefix, instead of
checking just the completely stripped-of-version link.

This make sure the right version of a dylib is linked, and allows
multiple copies of a library with different versions to be bundled and
linked.